move import queue to consensus-common (#1282)
