+++ core/client/src/client.rs
@@ -25,7 +25,7 @@ use runtime_primitives::{
 	Justification,
 	generic::{BlockId, SignedBlock},
 };
-use consensus::{ImportBlock, ImportResult, BlockOrigin};
+use consensus::{Error as ConsensusError, ErrorKind as ConsensusErrorKind, ImportBlock, ImportResult, BlockOrigin};
 use runtime_primitives::traits::{
 	Block as BlockT, Header as HeaderT, Zero, As, NumberFor, CurrentHeight, BlockNumberToHash,
 	ApiRef, ProvideRuntimeApi, Digest, DigestItem,
@@ -1019,7 +1019,7 @@ impl<B, E, Block, RA> consensus::BlockImport<Block> for Client<B, E, Block, RA> 
 	E: CallExecutor<Block, Blake2Hasher> + Clone + Send + Sync,
 	Block: BlockT<Hash=H256>,
 {
-	type Error = Error;
+	type Error = ConsensusError;
 
 	/// Import a checked and validated block. If a justification is provided in
 	/// `ImportBlock` then `finalized` *must* be true.
@@ -1044,9 +1044,10 @@ impl<B, E, Block, RA> consensus::BlockImport<Block> for Client<B, E, Block, RA> 
 
 		let parent_hash = header.parent_hash().clone();
 
-		match self.backend.blockchain().status(BlockId::Hash(parent_hash))? {
-			blockchain::BlockStatus::InChain => {},
-			blockchain::BlockStatus::Unknown => return Ok(ImportResult::UnknownParent),
+		match self.backend.blockchain().status(BlockId::Hash(parent_hash)) {
+			Ok(blockchain::BlockStatus::InChain) => {},
+			Ok(blockchain::BlockStatus::Unknown) => return Ok(ImportResult::UnknownParent),
+            Err(e) => return Err(ConsensusErrorKind::ClientImport(e.to_string()).into())
 		}
 
 		let import_headers = if post_digests.is_empty() {
@@ -1081,7 +1082,7 @@ impl<B, E, Block, RA> consensus::BlockImport<Block> for Client<B, E, Block, RA> 
 			"best" => ?hash,
 			"origin" => ?origin
 		);
-		result.map_err(|e| e.into())
+		result.map_err(|e| ConsensusErrorKind::ClientImport(e.to_string()).into())
 	}
 }
