+++ core/consensus/common/Cargo.toml
@@ -5,6 +5,8 @@ authors = ["Parity Technologies <admin@parity.io>"]
 description = "Common utilities for substrate consensus"
 
 [dependencies]
+log = "0.4"
+parking_lot = "0.7"
 substrate-primitives = { path= "../../primitives" }
 error-chain = "0.12"
 futures = "0.1"
@@ -13,3 +15,6 @@ sr-primitives = { path = "../../sr-primitives" }
 tokio = "0.1.7"
 parity-codec = "2.1"
 parity-codec-derive = "2.0"
+
+[dev-dependencies]
+substrate-test-client = { path = "../../test-client" }
