+++ core/consensus/common/src/lib.rs
@@ -24,8 +24,11 @@
 
 extern crate substrate_primitives as primitives;
 extern crate futures;
+extern crate parking_lot;
 extern crate sr_version as runtime_version;
 extern crate sr_primitives as runtime_primitives;
+#[cfg(any(test, feature = "test-helpers"))]
+extern crate substrate_test_client as test_client;
 extern crate tokio;
 
 extern crate parity_codec as codec;
@@ -33,6 +36,7 @@ extern crate parity_codec_derive;
 
 #[macro_use]
 extern crate error_chain;
+#[macro_use] extern crate log;
 
 use std::sync::Arc;
 
@@ -44,6 +48,7 @@ use futures::prelude::*;
 pub mod offline_tracker;
 pub mod error;
 mod block_import;
+pub mod import_queue;
 pub mod evaluation;
 
 // block size limit.
