+++ core/network/src/lib.rs
@@ -55,7 +55,6 @@ mod chain;
 mod blocks;
 mod on_demand;
 pub mod config;
-pub mod import_queue;
 pub mod consensus_gossip;
 pub mod error;
 pub mod message;
